Move some code back - scaled back Shift class
