+++ test/shift_test.rb
@@ -24,29 +24,4 @@ class ShiftTest < MiniTest::Test
     shift2 = Shift.new
     assert_equal Time.now.strftime("%d%m%y"), shift2.date
   end
-
-  def test_it_can_generate_keys
-    shift = Shift.new
-    key = "02715"
-    require "pry"; binding.pry
-
-    assert_equal ({
-      a: 2,
-      b: 27,
-      c: 71,
-      d: 15
-    }), shift.generate_keys(key)
-  end
-
-  def test_it_can_generate_offsets
-    shift = Shift.new
-    date = "040895"
-
-    assert_equal ({
-      a: 1,
-      b: 0,
-      c: 2,
-      d: 5
-      }), shift.generate_offsets(date)
-  end
 end
